+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ### 2.0.1-SNAPSHOT
+- Fix AIOOB exception due to incorrect Y position of fake chunk sections (#27)
 - Fix chunk light not properly unloading in tall worlds (#28)
 - Fix chunks leaking if client view distance is not greater than server view distance